requirements.txt
pytest-playwright
debugprint~=0.1.10
fastapi
python-dotenv